don't select empty region in backward direction
The call ‘(forward-char -1)’ makes sure that when target char is found
immediately before point we don't select empty region, but skip the char
and search further. Empty region to operate on is rarely what user
wants.

Note that this trick is already used in existing code, but only when
searching in forward direction. This commit adds the same behavior for
searching in backward direction too.